New event handlers - HELO/EHLO and AUTH
Added HELO and AUTH event handlers.

Minimum time between AUTH is at least 5 seconds. If a host sends
multiple AUTHs and the time between 2 AUTHs is less than 5 seconds, that
host is blacklisted.

For HELO/EHLO, if a host sends an IP address after the HELO/EHLO, that
host is blacklisted as well.

Added a simple application to visually test that AUTH works.